Fix x-default language tag description
The previous text did not fully align with the XMP spec. Why XMP has decided to (a) duplicate content in order to specify a default language, (b) not use BCP 47's `i-default`, and (c) have the redundant definition of both "first is default" and "`x-default` (which must be first) is default" I don't know, but we should align to it.